Refactor sidebar management and layout components
- Updated `useSidebarManager` to use `useXgenLayout` instead of `usePagesLayout`.
- Changed sidebar navigation links in `AdminSidebar` and `ManagerSidebar` to point to `/main` instead of `/chat`.
- Refactored `Monitor` and `TesterLogs` components to utilize `useXgenLayout`.
- Removed the old `XgenSidebar` component and replaced it with a new implementation in `sidebar/XgenSidebar.tsx`.
- Created a new sidebar configuration file `sidebar/sidebarConfig.ts` for managing sidebar items.
- Added new styles for the sidebar in `sidebar/Sidebar.module.scss`.
- Updated types for sidebar items in `sidebar/index.ts`.
