+++ src/app/main/sidebar/Sidebar.module.scss
