+++ src/app/main/sidebar/sidebarConfig.ts
