+++ src/app/main/workflowSection/types/index.ts
@@ -1,27 +1,5 @@
 import { ReactNode } from 'react';
 
-// 사이드바 관련 타입들
-export interface SidebarItem {
-    id: string;
-    title: string;
-    description: string;
-    icon: ReactNode;
-}
-
-export interface SidebarProps {
-    isOpen: boolean;
-    onToggle: () => void;
-    workflowItems?: SidebarItem[];
-    chatItems?: SidebarItem[];
-    trainItem?: SidebarItem[];
-    activeItem: string;
-    onItemClick: (itemId: string) => void;
-    className?: string;
-    initialChatExpanded?: boolean;
-    initialWorkflowExpanded?: boolean;
-    initialTrainExpanded?: boolean;
-}
-
 export interface ContentAreaProps {
     title: string;
     description: string;
